% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_func.R
\name{temp_func}
\alias{temp_func}
\title{Temporary function for R folder}
\usage{
temp_func(x)
}
\arguments{
\item{x}{Variable.}
}
\value{
No output, called for printing side-effect.
}
\description{
Temporary function for R folder
}
\examples{
temp_func("a")
}
