% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transform.epi_linelist}
\alias{transform.epi_linelist}
\title{Transform line list data}
\usage{
\method{transform}{epi_linelist}(`_data`, transformation = c("aggregate"), ...)
}
\arguments{
\item{...}{[dots] Extra arguments to be passed to specific methods.}

\item{`_data`}{A `<linelist>` object (sub class of `<data.frame>`)
output from [import()].}
}
\value{
An \R object
}
\description{
Transform line list data
}
\details{
The use of non-syntactic variable name for the first argument is because we
use the [transform()] generic from base R.
}
